move pool helper to pool.helper.ts
